+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/capture/SafariScreenshotImageProvider.java
@@ -139,6 +139,7 @@ public class SafariScreenshotImageProvider extends MobileScreenshotImageProvider
         devicesRegions.put(new RectangleSize(1170, 2532), new Rectangle[]{new Rectangle(0, 293, 1170, 1993)});
 
         devicesRegions.put(new RectangleSize(1640, 2360), new Rectangle[]{new Rectangle(0, 149, 1640, 2211)});
+        devicesRegions.put(new RectangleSize(1668, 2388), new Rectangle[]{new Rectangle(0, 149, 1640, 2239)});
     }
 
 }